# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Midpilot Connector Generator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/discovery/{session_id}/discovery": {
      "post": {
        "tags": [
          "Discovery"
        ],
        "summary": "Discover candidate documentation URLs",
        "description": "Enqueue a job to discover candidate documentation URLs for the given application.\nThe discovered URLs will be stored in the session.",
        "operationId": "discover_candidate_links_api_v1_discovery__session_id__discovery_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Session ID",
              "title": "Session Id"
            },
            "description": "Session ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateLinksInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Discovery"
        ],
        "summary": "Get discovery job status",
        "description": "Get the status of candidate links discovery job.",
        "operationId": "get_discovery_status_api_v1_discovery__session_id__discovery_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Session ID",
              "title": "Session Id"
            },
            "description": "Session ID"
          },
          {
            "name": "jobId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Job ID (optional)",
              "title": "Jobid"
            },
            "description": "Job ID (optional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusStageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseProgress": {
        "properties": {
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage",
            "description": "High-level stage, e.g., running, finished"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-friendly note about current work"
          }
        },
        "type": "object",
        "title": "BaseProgress"
      },
...
      "CandidateLinksInput": {
        "properties": {
          "applicationName": {
            "type": "string",
            "title": "Applicationname",
            "description": "Target application name"
          },
          "applicationVersion": {
            "type": "string",
            "title": "Applicationversion",
            "description": "Optional version string",
            "default": "latest"
          },
          "integrationType": {
            "type": "string",
            "enum": [
              "scim",
              "rest",
              "dummy"
            ],
            "title": "Integrationtype",
            "description": "Discovery protocol priority: scim, rest, or dummy.",
            "default": "dummy"
          },
          "llmGeneratedSearchQuery": {
            "type": "boolean",
            "title": "Llmgeneratedsearchquery",
            "description": "Use LLM to generate web search queries (default to use templates)",
            "default": false
          },
          "skipCache": {
            "type": "boolean",
            "title": "Skipcache",
            "description": "Whether to skip cached discovery output when input is the same.",
            "default": false
          },
          "enableLinkFiltering": {
            "type": "boolean",
            "title": "Enablelinkfiltering",
            "description": "Enable LLM-based filtering of irrelevant links",
            "default": true
          },
          "enableLinkRanking": {
            "type": "boolean",
            "title": "Enablelinkranking",
            "description": "Enable LLM-based ranking of candidate links",
            "default": true
          },
          "numQueries": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Numqueries",
            "description": "How many distinct search queries to run during discovery (allowed range: 1-8).",
            "default": 8
          },
          "maxResultsPerQuery": {
            "type": "integer",
            "title": "Maxresultsperquery",
            "description": "How many results to fetch per query (per backend)",
            "default": 10
          },
          "maxCandidateLinks": {
            "type": "integer",
            "title": "Maxcandidatelinks",
            "description": "Max number of candidate links to return after ranking/selection",
            "default": 10
          },
          "maxFilterLlmCalls": {
            "type": "integer",
            "title": "Maxfilterllmcalls",
            "description": "Maximum number of LLM calls for filtering irrelevant links",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "applicationName"
        ],
        "title": "CandidateLinksInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreateResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "title": "Jobid",
            "description": "Unique identifier of the created job."
          }
        },
        "type": "object",
        "required": [
          "jobId"
        ],
        "title": "JobCreateResponse"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "finished",
          "failed",
          "not_found"
        ],
        "title": "JobStatus"
      },
...
      "JobStatusStageResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "title": "Jobid",
            "description": "Job identifier"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current status of the job"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat",
            "description": "Job creation time (ISO 8601)"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat",
            "description": "Job start time (ISO 8601)"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat",
            "description": "Last update time (ISO 8601)"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result",
            "description": "Result payload when status is 'finished'"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Structured list of error lines/messages for pretty presentation. Each list item is a single error line. Backward-compatible with 'error' string."
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BaseProgress"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stage + message only progress info"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "status"
        ],
        "title": "JobStatusStageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API key validated and forwarded by Gravitee; backend access must be restricted to the gateway.",
        "in": "header",
        "name": "X-Gravitee-Api-Key"
      }
    }
  }
}